API application design for HomePage
New application interface using HTML and CSS.
All of the typos are fixed.
The RecipeSearch page does not exit anymore.
Incorporated a new search bar for HomePage to search for recipes.
Added a new page named "FavouritesRecipes".
The "FavouritesRecipes" page is still work in progress,
no connection between it and the API, so you will not be able to see
your favourite recipes yet.
The reference for "RecipeSearch.css" was deleted
the file did not exist in the project files.